requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
lightgbm==4.0.0
pyspark==3.4.1
mlflow==2.7.1
evidently==0.4.5

# API and web
fastapi==0.103.1
uvicorn==0.23.2
streamlit==1.26.0

# AWS
boto3==1.28.40
awscli==1.29.40

# Infrastructure
terraform-provider-aws==5.0.0

# Monitoring
prometheus-client==0.17.1
grafana-api-client==2.3.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# LangChain (optional)
langchain==0.0.267
openai==0.28.0
litellm==0.1.578

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1